# contracts/camera-stream.openapi.yaml
openapi: 3.1.0
info:
  title: Camera Stream ↔ AI Vision Contract
  version: 1.0.0
  description: API contract cho việc phân tích frame camera sử dụng AI (Team A2 - Camera Stream)
servers:
  - url: http://localhost:8000
    description: Camera Stream Service
  - url: http://localhost:4010
    description: Prism Mock Server
tags:
  - name: health
    description: Health check service
  - name: detection
    description: Camera frame analysis and tracking

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    AnalysisType:
      type: string
      enum: [PERSON_DETECTION, VEHICLE_DETECTION, UNKNOWN]
    Status:
      type: string
      enum: [PROCESSING, COMPLETED, FAILED]
    BoundingBox:
      type: object
      properties:
        x: { type: integer }
        y: { type: integer }
        width: { type: integer }
        height: { type: integer }
    DetectionResult:
      type: object
      required: [detectionId, status]
      properties:
        detectionId: { type: string, format: uuid }
        status: { $ref: '#/components/schemas/Status' }
        detectionType: { type: string }
        confidence: { type: number, minimum: 0, maximum: 1 }
        boundingBox: { $ref: '#/components/schemas/BoundingBox' }
        trackingId: { type: string }

security:
  - bearerAuth: []

paths:
  /health:
    get:
      tags: [health]
      summary: Kiểm tra service sẵn sàng
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: { type: string, example: "ok" }

  /detect:
    post:
      tags: [detection]
      summary: Phân tích frame camera (Async)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [cameraId, frameUrl, timestamp, requestId, analysisType]
              properties:
                cameraId: { type: string, example: "CAM-01" }
                frameUrl: { type: string, format: uri }
                timestamp: { type: string, format: date-time }
                requestId: { type: string }
                analysisType: { $ref: '#/components/schemas/AnalysisType' }
      responses:
        '202':
          description: Accepted - Request đang được xử lý
          content:
            application/json:
              schema:
                type: object
                properties:
                  detectionId: { type: string, format: uuid }
                  status: { $ref: '#/components/schemas/Status' }

  /detections:
    get:
      tags: [detection]
      summary: Danh sách detection (cursor pagination)
      parameters:
        - name: cursor
          in: query
          schema: { type: string, nullable: true }
        - name: limit
          in: query
          schema: { type: integer, default: 20 }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/DetectionResult' }
                  nextCursor: { type: string, nullable: true }
                  hasMore: { type: boolean }

  /detections/{detectionId}:
    get:
      tags: [detection]
      summary: Chi tiết detection
      parameters:
        - name: detectionId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DetectionResult' }